compile text.

+++ Sources/Elastos/Frameworks/Droid/Base/Core/src/elastos/droid/text/BoringLayout.cpp
@@ -225,7 +225,7 @@ ECode BoringLayout::Init(
 
     mPaint = IPaint::Probe(paint);
 
-    Float mtbottom, mttop, mtdescent, mtascent;
+    Int32 mtbottom, mttop, mtdescent, mtascent;
     Int32 mtwidth;
     metrics->GetBottom(&mtbottom);
     metrics->GetTop(&mttop);
